Move build_Uvqc
