+++ src/qforte/abc/uccpqeabc.py
@@ -11,7 +11,7 @@ from qforte.utils.trotterization import trotterize
 
 import numpy as np
 
-class UCCPQE(PQE):
+class UCCPQE(PQE, UCC):
 
     def fill_pool(self):
 
@@ -34,27 +34,6 @@ class UCCPQE(PQE):
         self._curr_grad_norm = 0.0
 
         self._Nm = [len(operator.jw_transform().terms()) for _, operator in self._pool_obj.terms()]
-
-    # TODO (opt major): write a C function that prepares this super efficiently
-    def build_Uvqc(self, amplitudes=None):
-        """ This function returns the QuantumCircuit object built
-        from the appropiate amplitudes (tops)
-
-        Parameters
-        ----------
-        amplitudes : list
-            A list of parameters that define the variational degrees of freedom in
-            the state preparation circuit Uvqc. This is needed for the scipy minimizer.
-        """
-
-        ansatz = UCC(self._trotter_number, self._tamps, self._tops, self._pool)
-        U = ansatz.ansatz_circuit(amplitudes)
-
-        Uvqc = qforte.QuantumCircuit()
-        Uvqc.add_circuit(self._Uprep)
-        Uvqc.add_circuit(U)
-
-        return Uvqc
 
     def verify_required_UCCPQE_attributes(self):
